# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MySchool Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /fees:
    get:
      operationId: listFees
      tags: [fees]
      summary: List fee records
      security:
        - BearerAuth: []
      parameters:
        - name: schoolId
          in: query
          schema:
            type: integer
        - name: studentId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, paid, overdue]
      responses:
        "200":
          description: Fee records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeListResponse"

    post:
      operationId: createFee
      tags: [fees]
      summary: Create fee record
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFeeRequest"
      responses:
        "201":
          description: Fee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fee"
  /fees/{feeId}/pay:
    post:
      operationId: payFee
      tags: [fees]
      summary: Mark fee as paid
      security:
        - BearerAuth: []
      parameters:
        - name: feeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Fee paid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fee"
  /notices:
    get:
      operationId: listNotices
      tags: [notices]
      summary: List notices
      security:
        - BearerAuth: []
      parameters:
        - name: schoolId
          in: query
          schema:
            type: integer
        - name: classId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Notices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoticeListResponse"

    post:
      operationId: createNotice
      tags: [notices]
      summary: Post a notice
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoticeRequest"
      responses:
        "201":
          description: Notice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notice"
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List teacher job postings
      parameters:
        - name: schoolId
          in: query
          schema:
            type: integer
        - name: subject
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Job postings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"

    post:
      operationId: createJob
      tags: [jobs]
      summary: Post a job opening
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
      responses:
        "201":
          description: Job posted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{jobId}/apply:
    post:
      operationId: applyJob
      tags: [jobs]
      summary: Apply to a job
      security:
        - BearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobApplicationRequest"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobApplication"
  /jobs/applications:
    get:
      operationId: listJobApplications
      tags: [jobs]
      summary: List job applications
      security:
        - BearerAuth: []
      parameters:
        - name: jobId
          in: query
          schema:
            type: integer
        - name: applicantId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Applications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobApplicationListResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Fee:
      type: object
      properties:
        id:
          type: integer
        schoolId:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        className:
          type: string
        amount:
          type: number
        feeType:
          type: string
        dueDate:
          type: string
          format: date
        paidDate:
          type: string
          format: date
        status:
          type: string
          enum: [pending, paid, overdue]
        description:
          type: string
      required:
        - id
        - schoolId
        - studentId
        - amount
        - status
    CreateFeeRequest:
      type: object
      properties:
        schoolId:
          type: integer
        studentId:
          type: integer
        amount:
          type: number
        feeType:
          type: string
        dueDate:
          type: string
          format: date
        description:
          type: string
      required:
        - schoolId
        - studentId
        - amount
        - feeType
        - dueDate
    FeeListResponse:
      type: object
      properties:
        fees:
          type: array
          items:
            $ref: "#/components/schemas/Fee"
        total:
          type: integer
        totalPending:
          type: number
        totalPaid:
          type: number
      required:
        - fees
        - total
    Notice:
      type: object
      properties:
        id:
          type: integer
        schoolId:
          type: integer
        classId:
          type: integer
        title:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [general, exam, holiday, ptm, event, urgent]
        postedBy:
          type: integer
        postedByName:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - schoolId
        - title
        - content
        - type
    CreateNoticeRequest:
      type: object
      properties:
        schoolId:
          type: integer
        classId:
          type: integer
        title:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [general, exam, holiday, ptm, event, urgent]
      required:
        - schoolId
        - title
        - content
        - type
    NoticeListResponse:
      type: object
      properties:
        notices:
          type: array
          items:
            $ref: "#/components/schemas/Notice"
        total:
          type: integer
      required:
        - notices
        - total
...
    Job:
      type: object
      properties:
        id:
          type: integer
        schoolId:
          type: integer
        schoolName:
          type: string
        schoolCity:
          type: string
        schoolLogoUrl:
          type: string
        title:
          type: string
        subject:
          type: string
        description:
          type: string
        requirements:
          type: string
        salary:
          type: string
        jobType:
          type: string
          enum: [full_time, part_time, contract]
        experienceRequired:
          type: integer
        location:
          type: string
        deadline:
          type: string
          format: date
        isActive:
          type: boolean
        applicationCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - schoolId
        - title
        - subject
        - isActive
    CreateJobRequest:
      type: object
      properties:
        schoolId:
          type: integer
        title:
          type: string
        subject:
          type: string
        description:
          type: string
        requirements:
          type: string
        salary:
          type: string
        jobType:
          type: string
          enum: [full_time, part_time, contract]
        experienceRequired:
          type: integer
        location:
          type: string
        deadline:
          type: string
          format: date
      required:
        - schoolId
        - title
        - subject
        - jobType
    JobListResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
      required:
        - jobs
        - total
    JobApplication:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        applicantId:
          type: integer
        applicantName:
          type: string
        coverLetter:
          type: string
        resumeUrl:
          type: string
        status:
          type: string
          enum: [pending, reviewed, shortlisted, rejected]
        appliedAt:
          type: string
          format: date-time
      required:
        - id
        - jobId
        - applicantId
        - status
    JobApplicationRequest:
      type: object
      properties:
        coverLetter:
          type: string
        resumeUrl:
          type: string
      required:
        - coverLetter
    JobApplicationListResponse:
      type: object
      properties:
        applications:
          type: array
          items:
            $ref: "#/components/schemas/JobApplication"
        total:
          type: integer
      required:
        - applications
        - total
...